requirements.txt
asgiref==3.8.1
blinker==1.9.0
boto3==1.35.92
botocore==1.35.92
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
dnspython==2.7.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
